Build: Update detekt to 1.17.1
With this update, the test sources below, which were not included by
default before, are now included by default:

- src/test/java
- src/test/kotlin

As such, for all the above test sources, Detekt has started to warn
about current violations based on the existing rules. As such, those
need to be resolved/suppressed accordingly. This is going to be done in
subsequent commits.