Build: Update detekt to 1.19.0
With this update, the configuration for the rules below got
updated. Also, one rule, specific to Jetpack Compose, was added:

Rules that got updated:
- LongParameterList (complexity)
- ForbiddenComment (style)

Rules that got added:
- FunctionNaming (naming)

As such, for all the above listed rules, Detekt has started to warn
about current violations based on these rule updates. As such those
rules need to be updated/added accordingly. This is going to be done in
subsequent commits.