Analysis: Suppress inconsistent arrays lint issue for available langs
This is suppressed as it causes the 'lintJetpackVanillaRelease' check to
fail.

This is because of this f281a2cbd365752cad1387001e42c78c9bbf7725
update translations commit, which added the new Jetpack related
'available_languages.xml' file, which has 17 number of languages,
compared to the original WordPress related such file, which has 51
instead. However, this is expected for Jetpack and as such suppressed.

Explanation: "When an array is translated in a different locale, it
should normally have the same number of elements as the original array.
When adding or removing elements to an array, it is easy to forget to
update all the locales, and this lint warning finds inconsistencies like
these.

Note however that there may be cases where you really want to declare a
different number of array items in each configuration (for example where
the array represents available options, and those options differ for
different layout orientations and so on), so use your own judgment to
decide if this is really an error."