Analysis: Suppress manifest resource lint issues for android manifest
This is suppressed as it causes the 'lintJetpackVanillaRelease' check to
fail.

This is because of this 5593a9152be83339ea5b10596464a00578abc611
update translations fallbacks for Jetpack commit, which added new
Jetpack related 'strings.xml' files, each with the 'widgets_enabled'
string in it (set to 'false').

Explanation: "Elements in the manifest can reference resources, but
those resources cannot vary across configurations (except as a special
case, by version, and except for a few specific package attributes such
as the application title and icon)."