Analysis: Update long parameter list ignore annotated configuration
With the '1.19.0' Detekt update, the Detekt 'ignoreAnnotated' feature
got updated. With that update, the 'javax.inject.Inject' configuration
was no longer working and had to be update to its simpler 'Inject'
version, which doesn't require the full import naming convention.

Also, as part of this update the 'javax.annotation.Generated' or
'Generated' configuration got removed as it is no longer (or was never)
necessary. Thus, making this overall configuration much more
straightforward.