Build: Update detekt to 1.20.0
With this update, the configuration for the ruleset below got
updated. Also, a couple of rules got fixed:

Ruleset that got updated:
- formatting

Rules that got fixed:
- TopLevelPropertyNaming (naming)
- UnnecessaryAbstractClass (style)

As such, for the above listed ruleset and all the above listed rules,
Detekt has started to warn about current violations based on these
updates. As such those need to be updated accordingly. This is going to
be done in subsequent commits.