Analysis: Remove formatting section and depend on the formatting config
With the '1.20.0' Detekt update, the Detekt 'formatting' section got
moved out of the default config and into its own 'formatting' related
configuration. This is done because the 'formatting' ruleset provided by
Detekt is updated to use a new mechanism that allow authors to define
the default configuration for their custom rules, which will then be
merged together with the user configuration and can be overridden by the
user if they wish.

Formatting Config: https://github.com/detekt/detekt/blob/main/
detekt-formatting/src/main/resources/config/config.yml

Actually, if this is not done, then, when running the './gradlew detekt'
task it fails with the below message:

"* What went wrong:
Execution failed for task ':WordPress:detekt'.
> Run failed with 1 invalid config property.
        - Property 'formatting' is misspelled or does not exist."