Build: Update detekt to 1.21.0
With this update, some rules, like the below, which were disabled
by default, are now enabled by default instead:

- UseCheckOrError (style)
- ExplicitItLambdaParameter (style)
- UseRequire (style)
- DestructuringDeclarationWithTooManyEntries (style)

Also, the below rules got fixed:
- ReturnCount (style)

As such, for all the above listed rules, Detekt has started to warn
about current violations based on these rules. As such those need to
be resolved/suppressed accordingly. This is going to be done in
subsequent commits.