Analysis: Add back the default todo value for the forbidden comment rule
Updating this rule was discussed in the link shared below and thus this
is a good opportunity to fine-tune Detekt to update it:

Link: https://github.com/wordpress-mobile/WordPress-Android/issues/14011
#issuecomment-775885924

Also, as part of enabling this rule, some warnings based on that was
suppressed (them being TODOs), that is, instead of being resolved since
a resolution would require a proper solution (and testing).

PS: This commit align WPAndroid with WCAndroid in terms of Detekt rules.