Analysis: Enable ignore default parameters for the long param list rule
Updating this rule to align WPAndroid with WCAndroid in terms of Detekt
rules.

As part of updating this rule you will notice that all such baseline
rules got removed and either added explicitly, inline/close to source or
just deleted since they are no longer necessary. Thus, you will notice
some new suppress annotations that got added as part of this commit.
But, all those warnings are being added because they got moved from the
baseline and not because they are new.