Analysis: Resolve utility class with public constr. warns for fluxc mod.
Description: "A class which only contains utility variables and
functions with no concrete implementation can be refactored into an
'object' or a class with a non-public constructor. Furthermore, this
rule reports utility classes which are not final."

For more info see: https://detekt.dev/docs/rules/style/
#utilityclasswithpublicconstructor

Note: As part of this commit, the unused 'versionRegex' was also
removed from the 'WhatsNewAppVersionUtils' utility class.