Bump `versionCode` manually to adjust for the zalpha removal
This manual bump of `versionCode=1260` was done so that developers would not experiment installation issues due to us removing the `zalpha` flavor and moving other flavors away from using the `alpha.versionCode` property (which we just deleted) in favor of the `versionCode` property.

Note that, because that change was done manually and without doing a beta to PlayStore alongside it, the `versionCode` updated in this commit does *not* reflect a `versionCode` last used in the PlayStore for the current `versionName`. That is, this commit temporarily make those properties not match the versions last uploaded in the PlayStore anymore (which were `versionCode=1259` for that `versionName=20.5-rc-1`).

This disconnection (with the values actually used for uploaded app bundled to PlayStore) was deemed acceptable though, to ease the transition made by removing the `zalpha` flavor and `alpha.version{Name,Code}` properties — See discussion in https://github.com/wordpress-mobile/WordPress-Android/pull/17023#issuecomment-1209943410

This is a transcient state anyway, as the `versionName` and `versionCode` properties will go back to match the values for builds uploaded in PlayStore as soon as we do the next beta or final build.