Analysis: Suppress spread operator warnings for fluxc module
Description: "In most cases using a spread operator causes a full copy
of the array to be created before calling a method. This has a very high
performance penalty. Benchmarks showing this performance penalty can be
seen here: https://sites.google.com/a/athaydes.com/renato-athaydes/
posts/kotlinshiddencosts-benchmarks

The Kotlin compiler since v1.1.60 has an optimization that skips the
array copy when an array constructor function is used to create the
arguments that are passed to the vararg parameter. When type resolution
is enabled in detekt this case will not be flagged by the rule since it
doesn't suffer the performance penalty of an array copy."

For more info see: https://detekt.dev/docs/rules/performance
#spreadoperator