Add code to ensure locales declaration consistency
By comparing the locales limited by the `resourcesConfiguration` field in a product flavor with the ones we hardcoded in our Fastlane scripts.

This way if we decide to update either place in the future (e.g. add new locales to download from GlotPress for Jetpack by updating the `JP_APP_LOCALES` constant in `fastlane/lanes/localization.rb`) we'll be reminded to ensure we updated the other as well (i.e. the `resourceConfigurations` in the `jetpack` flavor of the `WordPress/build.gradle` file)