Analysis: Suppress data cass should be immutable warn for comments dao
This is done because otherwise the 'UnifiedCommentLeveler' class on the
WPAndroid side is not compiling. Actually, not compiling is not an issue
as this change is expected to break clients. But, in this case to fix
this logic would require too much work, or might not even be possible on
the current solution. Thus, suppressing 'DataClassShouldBeImmutable', at
least for now, in order to not block using any new versions of FluxC, is
the recommended action.

Description: "Data classes should mainly be used to store immutable
data. This rule assumes that they should not contain any mutable
properties."

For more info see: https://detekt.dev/docs/rules/style/
#dataclassshouldbeimmutable