Deps: Update fluxc to latest trunk hash
This 'FluxC' hash updates 'FluxC' to that latest 'trunk' version where
the 'RefreshSitesXMLRPCPayload' data class becomes immutable.

As such, the [refreshXmlRpcSites] method for the
[LoginUsernamePasswordFragment] class needs updating to adhere to this
change.

This step is required in order to avoid the client apps like 'WPAndroid'
and 'WCAndroid' crashing while using a newer version of 'FluxC' on their
side due to transitive dependencies taking precedence and using the
latest 'FluxC' version, the one defined on the client's side.