Deps: Update fluxc to latest trunk hash
This 'FluxC' hash updates 'FluxC' to that latest 'trunk' version where
the 'RefreshSitesXMLRPCPayload' data class becomes immutable. The
'Login' library and its 'FluxC' dependency is using this model and as
such requires this change.

For more info see: https://github.com/wordpress-mobile/
WordPress-Login-Flow-Android/pull/92

This, in coordination to the previous 'Login' update commit that
followed, is done in order to make sure that both, this WPAndroid client
and its 'Login' library dependency, are pointing to that 'FluxC' version
where everything works as expected, without crashes or weird behaviors.