Add the OLD `play_store_desc` copy under a new key
This is only to help polyglots translate the new and shorter copy for the `play_store_desc` key.

By re-introducing the old copy under a new `play_store_desc_old_copy_do_not_translate` key, this will allow GlotPress to see the source of this new copy and suggest the translation memory for it, which we hope will help polyglots use their previous translation for this previous copy as a baseline to translate the new, shorter copy.

Because the new copy is very similar to the old one, with many paragraphs unchanged and only a couple of sentences removed, allowing polyglots to see the translations they used to provide for the old copy (even if that past translation they provided was over the limit of 4000 characters back then) should help them in translating the new one, by copy/pasting parts of the translation memory for the old one and use it as a baseline, instead of having to translate the new copy from scratch.