Update WPLocaleTestRule + chain it with fastlane's
The latest version of fastlane Screengrab lib does not pass the `endingLocale` when triggering the instrumentation tests, and that fastlane's own built-in `LocaleTestRule` implementation has changed to their latest way to do it, so we need to update our custom `WPLocaleTestRule` (which is based on the model of fastlane's `LocaleTestRule`) accordingly.

Then, to ensure that both the code from fastlane's own `LocaleTestRule` (which consists of changing the device locale but also calling `Screengrab.setLocale` and setting internal stuff for screengrab to work properly and as expected) and our own code from `WPLocaleTestRule` (which handles our WP-specific, in-app locale switching logic) are run and are executed in the right order (to ensure that both fastlane's rule and expectations are respected and that our custom logic is handled), we need to use a `RuleChain` on each of our screenshot tests (`{WP,JP}ScreenshotTest.java`)

---

Note that if I'm not mistaken, our `WPLocaleTestRule` is not really necessary when we're in the case where we run the screenshot tests using `fastlane screenshots` lane and start on a fresh and wiped-out emulator where the APK have never been installed, as in such case the in-app locale switcher would never have been changed by a user, and the User Preference for the custom locale would be `null`, and the app will then always use the Device Locale (which fastlane's `LocaleTestRule` takes care of changing) in such case. But I still think it's valuable to add our `WPLocaleTestRule` to the `RuleChain` for the case when developers work on fixing the screenshot UI tests and might try to run them from Android Studio, via an emulator they might not have wiped out first, and on which they might already have the APK installed from previous work, and in which they might have changed the in-app locale. So better safe than sorry :)