Lock LoginActivity to portrait mode
This is uesful to avoid orientation changes during the Jetpack
interactive landing screen animation, but also affects the WordPress
app. Note that this will be locked in portrait even when the feature
flag is not enabled. This should be ok, imo, since the UI in landscape
is not ideal anyway.