Extract physics model to view model and use CompositionLocal
This refactors the physics model to the view model for the fragment, and
also makes the consuming composables depend on position, rather than
acceleration. This avoids a duplication of state in the normal and
blurred composables, preventing the possibility that they can drift
"out-of-sync" with one another, since they previously each maintained
their own source of truth regarding the offset position.

This change also enables "donut hole skipping" while making use of
CompositionLocal for the position, allowing the intermediate composables
in the hierarchy to be completely decoupled from the position updates.