Build: Remove unnecessary source/target compatibility for java 7
This configuration was removed from both, the 'annotations' and
'processors' modules.

This configuration was removed as a follow-up on this discussion below
when its was discussed whether we should be removing the JVM
source/target explicitly at that point. After a short discussion, it
was decided to not do that back then, but nowadays there is no much
point in keeping this configuration any longer.

Discussion: https://github.com/wordpress-mobile/WordPress-iOS/pull/19335