Analysis: Resolve unnecessary safe call type warning
Warning Message: "Unnecessary safe call on a non-null receiver of type
CharSequence. This expression will have nullable type in future
releases"

The 'mPositiveButtonLabel' is being reverted back to being a nullable
field, that is, instead of it being a 'lateinit' non-null field.

PS: This field was nullable in the past, but this
24a053275df69f69924ab7dbf9ff60787ce30f8d commit made it into
'lateinit' non-null.