Analysis: Resolve opt-in experimental coroutines api warnings
1st Warning Messages: "This declaration is experimental and its usage
should be marked with '@kotlinx.coroutines.ExperimentalCoroutinesApi' or
'@OptIn(kotlinx.coroutines.ExperimentalCoroutinesApi::class)'"

Adding the '@OptIn(ExperimentalCoroutinesApi::class)' annotation on
these functions is the recommended action to resolve this 'GlobalScope'
warnings. However, this results to the below '@OptIn' warnings:

2nd Warning Message: "This annotation should be used with the
compiler argument '-Xopt-in=kotlin.RequiresOptIn'"

Adding the '-Xopt-in=kotlin.RequiresOptIn' free compiler argument is the
recommended action to resolve this '@OptIn' warning.