Fix: Resolve class cast exception on edittext preference with validation
After updating the 'android.preference.PreferenceManager' import to
'androidx.preference.PreferenceManager' the below
'java.lang.ClassCastException' started appearing on 'EditPostActivity':
org.wordpress.android.ui.prefs.EditTextPreferenceWithValidation
cannot be cast to androidx.preference.Preference

This crash happens because the [PreferenceManager.setDefaultValues(...)]
is instantiating an 'EditTextPreferenceWithValidation' preference, via
the 'R.xml.account_settings' preference file, which is a type of
'android.preference.EditTextPreference' and thus a type of
'android.preference.Preference'. Because 'android.preference.Preference'
is not an 'androidx.preference.Preference', as expected, a class cast
exception is thrown.

As such, reverting the 'androidx.preference.PreferenceManager' change is
the only logical solution, for now, and until a full migration of
'android.preference.Preference' to 'androidx.preference.Preference'
occurs.