Analysis: Resolve on lifecycle event deprecated warns for story media
Warning Message: "'OnLifecycleEvent' is deprecated. Deprecated in Java"

Replacing 'LifecycleObserver' with 'DefaultLifecycleObserver' and then
using this interface's functions via a public override instead of
decorating a function with the '@OnLifecycleEvent' annotation is now the
recommended and more explicit way forward.

PS: The previous annotation based solution required the usage of code
generation or reflection, which should be avoided.

For more info see:
- androidx.lifecycle.OnLifecycleEvent (Documentation):
https://developer.android.com/reference/androidx/lifecycle/
OnLifecycleEvent
- androidx.lifecycle.DefaultLifecycleObserver (Documentation):
https://developer.android.com/reference/androidx/lifecycle/
DefaultLifecycleObserver