Analysis: Suppress async task deprecated warnings
Warning Messages:
- "'AsyncTask<Params : Any!, Progress : Any!, Result : Any!>'
is deprecated. Deprecated in Java"
- "'THREAD_POOL_EXECUTOR: Executor!' is deprecated. Deprecated in Java"
- "'constructor AsyncTask<Params :
Any!, Progress : Any!, Result : Any!>()' is deprecated.
Deprecated in Java"
- "'executeOnExecutor(Executor!, vararg Void!):
AsyncTask<Void!, Void!, List<NoteBlock>?>!' is deprecated.
Deprecated in Java"

These deprecated warnings are suppressed, that is, instead of being
resolved, since a resolution would require a proper migration.

You will also notice that I suppressed the deprecated warning both, on
the file level and closer to the deprecation itself. In addition to
that, you will notice that I suppressed the deprecated warning for other
such warnings, like 'GeneratedNoteBlock'. I did that because suppressing
on the file level was necessary due to the 'android.os.AsyncTask' import
that can't be suppressed any other way. But, when doing so, every
deprecated warning is suppressed as well. At which point the deprecated
information is lost. This way, adding first the '@Suppress
("DEPRECATION")' closer to source and then adding it on the file level
makes sure that this information is not lost, at least for the existing
code.

Explanation: "This class was deprecated in API level 30. Use the
standard java.util.concurrent or Kotlin concurrency utilities instead."