Analysis: Suppress fragment manager deprecated warnings
Warning Messages:
- "'getter for fragmentManager: FragmentManager?' is deprecated
Deprecated in Java"
- "'requireFragmentManager(): FragmentManager' is deprecated.
Deprecated in Java"
- "'constructor FragmentPagerAdapter(FragmentManager)' is deprecated.
Deprecated in Java"
- "'show(FragmentManager!, String!): Unit' is deprecated.
Deprecated in Java"

These deprecated warnings are suppressed, that is, instead of being
resolved, since a resolution would require a proper migration.

You will also notice that I suppressed the deprecated warning both, on
the file level and closer to the deprecation itself. In addition to
that, you will notice that I suppressed the deprecated warning for other
such warnings, like 'DialogFragment'. I did that because suppressing
on the file level was necessary due to the 'android.app.FragmentManager'
import that can't be suppressed any other way. But, when doing so, every
deprecated warning is suppressed as well. At which point the deprecated
information is lost. This way, adding first the '@Suppress
("DEPRECATION")' closer to source and then adding it on the file level
makes sure that this information is not lost, at least for the existing
code.

For more info see:
- androidx.fragment.app.Fragment.getFragmentManager (Docs):
https://developer.android.com/reference/androidx/fragment/app/
Fragment#getFragmentManager()
- androidx.fragment.app.Fragment#.requireFragmentManager (Docs):
https://developer.android.com/reference/androidx/fragment/app/
Fragment#requireFragmentManager()