Remove DRIFT parameter from animation model
Originally, the DRIFT parameter was used to provide movement for cases
where the device lacked sensor support. This was implemented
incorrectly, since it was only ever applied in the `onSensorChanged`
function, which never runs on such devices. It also had a side effect
which was not part of the original designs: it biased the interaction
such that it resulted in movement when the device was flat on a table.

Since this was not ever a requirement, it may be simpler to just remove
this parameter, since we can provide default behavior for devices
without sensor support instead via default sensor values (in this case,
a default value for the device pitch).