Analysis: Suppress on activity for result deprecated warnings for java
This is an accompanying commit to a previous such commit, which was
mainly targeted Kotlin classes. As such, this commit now mainly targets
Java classes.

You will also notice that not all such Java classes are being annotated
with the '@SuppressWarnings("deprecation")'. This is due to the fact
that different versions of libraries are using different implementation
of 'onActivityResult(...)', which result to this suppression being
redundant if added, or at least seeming so. Thus, by adding that,
another set of such redundant suppression warnings are adding to the
list of overall warnings, and thus, at least for now, this is being
avoided in order to not having to then remove those again afterwards.

For more info see this commit: 103b90ccd3664c3295c772d2fcb4b119fc3750cd