Analysis: Suppress google app client & fb instance id deprecated warns
Warning Messages:
- "'GoogleApiClient' is deprecated. Deprecated in Java"
- "'Builder' is deprecated. Deprecated in Java"
- "'ConnectionCallbacks' is deprecated. Deprecated in Java"
- "'FirebaseInstanceId' is deprecated. Deprecated in Java"
- "'deleteInstanceId(): Unit' is deprecated. Deprecated in Java"

These deprecated warnings are suppressed, that is, instead of being
resolved, since a resolution would require a proper migration.

You will also notice that I suppressed the deprecated warning both, on
the file level and closer to the deprecation itself. I did that because
suppressing on the file level was necessary due to the
'com.google.android.gms.common.api.GoogleApiClient' and
'com.google.firebase.iid.FirebaseInstanceId' imports that can't be
suppressed any other way. But, when doing so, every deprecated warning
is suppressed as well. At which point the deprecated information is
lost. This way, adding first the '@Suppress("DEPRECATION")' closer to
source and then adding it on the file level makes sure that this
information is not lost, at least for the existing code.

For more info see:
- Compiler Warnings as Errors - WordPress Module - AppInitializer -
Google API Client & Firebase Instance Id #17229 (Issue:)
https://github.com/wordpress-mobile/WordPress-Android/issues/17229