Refactor: Remove unused get build config string/value functions
Both these functions are unused and/or contains compile warnings that
needs resolving.

Warning Messages:
- "Unsafe use of a nullable receiver of type Package?"
- "Type mismatch: inferred type is String? but String was expected"

Instead of resolving any compile warnings related to unused code, it is
better to remove such code and stop maintaining it altogether.

In the future, and if need be, this code can be added back, with an
explicit comment that it shouldn't be removed anymore, thus suppressing
it with an '@Suppress("unused")' annotation, along with a reasoning
behind keeping unused code like that.