Analysis: Suppress cast can never succeed cast warning
Warning Message: "This cast can never succeed"

This type warning is suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation and testing.
As such, it might be best to ignore this as out of scope, for now, and
so as to not introduce any breaking changes to this functionality
overall.