Analysis: Resolve non exhaustive when statements when warnings
Warning Message: "Non exhaustive 'when' statements on enum will be
prohibited in 1.7, add 'Xyz', 'Xyz' branch(es) or 'else' branch instead"

You will notice that I optimized the imports on all those when
statements and the imports they were using because one or no import is
always better than having multiple imports for no particular reason.
Also, having the extra information like 'FormattableRangeType.SITE'
instead of plain 'SITE' is usually better for readability purposes as
the reader will immediately understand that this 'SITE' is
'FormattableRangeType' related, without the need to navigate to the
'SITE' in order to understand when that is coming from.