Refactor: Replace run blocking test block with coroutines utils test
In addition to that, this commit also replaces the 'runBlockingTest'
block, from the 'kotlinx.coroutines.test' package, with the 'test'
block, from [CoroutinesUtils], that is using 'runBlocking' underneath
anyway. This was done to make tests simpler to reason about and  easier
to be migrated to Coroutines version 1.6, which completely changes how
testing with Coroutines works. The only tests that are still using
'runBlockingTest', and doing that purposefully are the below:
- [BatchModerateCommentsUseCaseTest]
- [ModerateCommentsWithUndoUseCaseTest]
- [PaginateCommentsUseCaseTest]
- [QRCodeAuthViewModelTest]
- [SnackbarSequencerConcurrentTest]
- [UnifiedCommentListViewModelTest]

Also, the 'coroutineScope' rule annotations, that were pointing to
[MainCoroutineScopeRule] were removed as well, that is, where they were
indeed unnecessary.