Analysis: Suppress unchecked cast test cast warnings
Warning Message: "Unchecked cast: Xyz to Xyz"

These type warnings are suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore those as out of scope, for now, and so as to not
introduce any breaking changes to these tests overall.