Analysis: Suppress cast can never succeed test cast warnings
Warning Message: "This cast can never succeed"

These type warnings are suppressed, that is, instead of being resolved,
since a resolution would require a proper investigation. As such, it
might be best to ignore those as out of scope, for now, and so as to not
introduce any breaking changes to these tests overall.