Analysis: Enable all warnings as errors for all modules
This 'allWarningsAsErrors' configuration was previously applied on the
module level in order to make sure that, as the overall 'Kotlin Warnings
as Errors' work is progressing, no new warnings are added to a
specific module, which is already free of warnings.

However, now that the overall 'Kotlin Warnings as Errors' work is
complete on all modules, this module level 'allWarningsAsErrors'
configuration is now replaced by a root level 'allWarningsAsErrors'
configuration that will be applied by default to all modules.