Add Parent Page item to Page Settings in Editor
Adding the section for Page Attributes and the Parent Page field as well inside
it. Right now the added changes also include exposing the parentId in the
EditPostRepository so we can show it in the item for now (ideally it will show
the Parent Page Title), but no interaction is programmed yet.

This commit also refactors a small reusable piece of logic used to hide some
UI elements in case of Pages vs Posts, so the logic was extracted to a method.

A unit test was added to EditPostRepository for the new exposed field.