Add parentTitle fetching to EditPostRepository
This exposes a new plain String field parentTitle which is fetched whenever
a post is loaded in the EditPostRepository as well as a method to update it
manually whenever the original post.parentId might have changed, such as after
a update or updateAsync call setting a new parentId.

Several other classes and unit tests needed to be updated to avoid breaking
the test suite / compilation, since a new parameter was needed in the
loadPostByLocalPostId method.