Get parentTitle from store only when needed (PR comment)
Improved the code to only fetch the parent title for the current parentId
when a caller asks for getParentTitle AND we know that the parentID changed,
storing the current <parentId, parentTitle> pair as a ParentPage internal
object in EditPostRepository.

This avoids introducing the siteModel in the loadPostByLocalPostId parameters
and makes more sense from a caller perspective (no need to know when to update
the title).