Build: (DV) Add dependency versions plugin
In a multi-project build, running this task in the root project will
generate a consolidated/merged report for dependency updates in all
subprojects. Alternatively, you can run the task separately in each
subproject to generate separate reports for each subproject.

As such, this plugin is not being applied on the root level, but
separately and per module, in order to generate a separate report for
each module, that is, instead of a merged report for the whole project.

For more info see: https://github.com/ben-manes/gradle-versions-plugin
#multi-project-build