Build: (DV) Configure dependency versions plugin
This commit configures the dependency version plugin in such a way as
to make sure that only stable versions are being included in the report
and advise for update. As such the 'dev', 'alpha', 'beta' and 'rc'
related version are being excluded from being valid version for update.

This is a very basic and generic configuration, other configurations
might exclude or include more versions. As such, consider this
configuration as the bare minimum and most generic, especially since
both 'checkConstraints' and 'checkBuildEnvironmentConstraints' are being
disabled.

For more info see: https://github.com/ben-manes/gradle-versions-plugin
#rejectversionsif-and-componentselection