Don't show local-only pages in Page Parent selection list
Even though the PageParent screen is not called when the device is offline,
it's better to put an additional safe-guard in case this page is opened when
there are pages that are local-only (created while device was offline and not
published remotely yet), as they cannot be used as parent to other pages. So
the changes here filter out local-only pages (the ones that have negative
remote IDs).