Fix potential crash on `download_metadata_strings`
If called with `skip_release_notes:true` and without a `version:` nor `skip_commit:true`, the commit messages would try to include the version name but crash if not provided like in such case.
So now we only include the `" for #{version}"` part in the commit message if such a version was provided.