Remove dependency on site id
This is possible now because we are preserving local ids. This means we
can rely on data relationships for the site entity to remain intact
after transering the db records.

This also allows us to be smarter about sending the list of post ids to
be transferred, as we will not need to do this site by site.