Implement processing state for error screen
Unfortunately the approach with having two separate data classes for each error type can't work together with a boolean for `isProcessing`, since we need to cast the type in order to be able emit only this change on the error state. That would require two separate casts for each error type resulting in overcomplicated code. I went forward with the fallback to a single data class that is be instantiated with different values for the two error types.