Analysis: Remove unnecessary commons text word utils capitalize
The 'WordUtils.capitalize(...)' can be safely replaced with Kotlin's
'StringsJVM.capitalize()' string extension function. Then, and since
this 'capitalize' is now deprecated it can be safely converted to the
newly recommended Kotlin's 'String.replaceFirstChar(...)' string
extension function.

FYI: The tests within the 'StatsDateFormatterTest' test suite verify
correctness of this change, especially the fact that the this test
succeeds: 'prints a month date in another language'