Refactor: Rename stats related date utils to stats date utils
This stats' related 'DateUtils' is not conflicting with the new created
and generic 'DateUtils' and as such it is better to name them difference
to avoid any future confusion on which utility to use and why. Renaming
the stats' such utility to 'StatsDateUtils' makes it explicit that this
utility should be used within stats.