Move the access token dispatch into the login helper
Though this emission does not signal failure (i.e. it can fail
silently), it can still be encapsulated within the login helper. Since
the token is persisted in a private shared preference, it should be less
error-prone. If it is necessary, later we can set the preference more
directly with `commit` instead of `apply`.