Deps: Remove wiremock mocks version
This '2.23.2' version of 'WireMock' specific to the 'mocks' module is a
discrepancy that should be aligned with the main 'WireMock' version used
on all other modules.

PS: Actually, this '2.23.2' version is anyway being updated to the main
'2.26.3' version due to transitive dependencies. As such, this
versioning related configuration change shouldn't matter much.