Build: Move included builds to fix dependabot
This change moves the configuration of included builds to a separate
file in order to fix Dependabot as it doesn't understand the content of,
for example 'ext.localFluxCPath'. Dependabot looks for known schemas of
version declaration on both, the 'build.gradle' and 'settings.gradle'
file.

As such moving this included builds specific configuration to a separate
file, with a different naming convention, will make Dependabot ignore
this file. As such, it will start resolving the dependencies again.

FYI: See Dependabot error below:

------------------------------------------------------------------------

Dependabot couldn't find a build.gradle.

Dependabot requires a build.gradle to evaluate your Java dependencies.
It had expected to find one at the path:
/ext.localFluxCPath/build.gradle.

If this isn't a Java project, you may wish to disable updates for it in
the .github/dependabot.yml config file in this repo.

------------------------------------------------------------------------