Deps: Remove unused androidx gridlayout dependency
This dependency was added as part of updating the 'Support Libraries' of
this module back in Feb 2019 (8f9d6b45ca69255cea38d64067e67faea113cb60).
However, this dependency is unused and 'GridLayoutManager' is the only
such related import found in this project. In addition to that, this is
specific to the 'RecyclerView' dependency and also found only on the
'WordPress' module.

As such, this leftover from back then can be now safely removed.