Deps: Remove unused androidx legacy support v4 dependency
This dependency was added as part the migration process to 'AndroidX'
for this module back in Jun 2019
(b7ac12d358f7238beb0ca48fc56916ff3c6b4128). However, this
dependency is unused and 'Widget Space' is the only such related
usage found in this project. In addition to that, this is specific to
the 'Legacy Support Core UI' dependency and only found on the
'WordPress' module.

As such, this leftover from back then can be now safely removed.