Deps: Remove unused jsoup dependency
This dependency was added as part of switching to latest 'Aztec' build
that wasn't passing on internal dependencies to this module back in
Aug 2017 (affc11fcafce31832261cca0c45d7fd885a96e60). However, this
dependency is unused and 'Jsoup', 'Document' and 'Element' is the only
such related imports found in this project. In addition to that, this is
specific to the 'WordPress' module.

It also seems that there is no longer such a requirement for those extra
Aztec dependencies as they seems to be included with the library and not
being stripped out, by Jitpack or otherwise. See related comment below,
which was included on top of the dependency that got deleted.

"Required Aztec dependencies (they should be included but Jitpack seems
to be stripping these out)"

As such, this leftover from back then can be now safely removed.