Deps: Remove unused androidx exif interface dependency
This dependency was added as part of updating the 'Support Libraries' of
this module back in Feb 2019 (8f9d6b45ca69255cea38d64067e67faea113cb60).
However, this dependency is unused and no related import found in this
project.

As such, this leftover from back then can be now safely removed.