Deps: Update androidx lifecycle to 2.5.1
Release Notes: https://developer.android.com/jetpack/androidx/releases/
lifecycle#2.5.1

------------------------------------------------------------------------

Note that the 'androidxComposeLifecycleVersion' version got also updated
to '2.5.1' since it is currently on the same versioning track as
'androidxLifecycleVersion' itself.

Lifeycle Artifacts: https://developer.android.com/jetpack/androidx/
releases/lifecycle

------------------------------------------------------------------------

As part of this update the '-Xjvm-default=all' Kotlin related free
compiler argument was added because otherwise
the 'StatsViewAllViewModelFactory' class was failing to compile due to
the below error:

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
> Task :WordPress:compileWordpressWasabiDebugKotlin FAILED
e: /Users/.../WordPress-Android/WordPress/src/main/java/org/wordpress
/android/ui/stats/refresh/StatsViewAllViewModelFactory.kt: (64, 7):
Inheritance from an interface with '@JvmDefault' members is only allowed
with -Xjvm-default option
- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

For more info see:
- Kotlin Blog: https://blog.jetbrains.com/kotlin/2020/07/
kotlin-1-4-m3-generating-default-methods-in-interfaces/
- Google Issue Tracker: https://issuetracker.google.com/issues/217593040
- Stack Overflow Answer (Ian Lake): https://stackoverflow.com/questions/
70992947/how-do-i-resolve-error-message-inheritance-from-an-interface-
with-jvmdefault

------------------------------------------------------------------------

Also, as part of this change, the below 'kotlinx.coroutines.test'
related imports had to be explicitly added, otherwise the corresponding
tests were failing to compile as those imports are now seen as missing:
- kotlinx.coroutines.test.advanceUntilIdle
- kotlinx.coroutines.test.advanceTimeBy
- kotlinx.coroutines.test.pauseDispatcher
- kotlinx.coroutines.test.resumeDispatcher

PS: 3/4 of the above imports are now deprecated as well. As such, those
need to be updated accordingly. This is going to be done in subsequent
commits. Below are the deprecated imports:
- kotlinx.coroutines.test.advanceTimeBy
- kotlinx.coroutines.test.pauseDispatcher
- kotlinx.coroutines.test.resumeDispatcher

In addition to that, lots of other 'kotlinx.coroutines.test' imports got
deprecated as well, like 'TestCoroutineDispatcher',
'TestCoroutineScope', 'runBlockingTest', etc. All those need to be
dealt with as well.

This probably happened due to a transitive dependency update on
'kotlinx.coroutines', which bumped Coroutines to 1.6.1 for the whole
project, that is, as part of the 'androidx.lifecycle' update to '2.5.1'.
For example see some results below from running this dependencies
'./gradlew WordPress:dependencies' task:

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
androidx.lifecycle:lifecycle-runtime-ktx:2.5.1
+--- androidx.annotation:annotation:1.0.0 -> 1.2.0
+--- androidx.lifecycle:lifecycle-runtime:2.5.1 (*)
+--- org.jetbrains.kotlin:kotlin-stdlib:1.6.21 (*)
\--- org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1 (*)
- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
androidx.lifecycle:lifecycle-livedata-ktx:2.5.1
+--- androidx.lifecycle:lifecycle-livedata:2.5.1 (*)
+--- androidx.lifecycle:lifecycle-livedata-core-ktx:2.5.1 (*)
+--- org.jetbrains.kotlin:kotlin-stdlib:1.6.21 (*)
\--- org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.1 (*)
- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
androidx.lifecycle:lifecycle-viewmodel-ktx:2.3.1 -> 2.5.1
+--- androidx.lifecycle:lifecycle-viewmodel:2.5.1 (*)
+--- org.jetbrains.kotlin:kotlin-stdlib:1.6.21 (*)
\--- org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.1 (*)

------------------------------------------------------------------------