Test: Migrate to new coroutine test rule
With this change, this 'CoroutineTestRule' is also moved to
'BaseUnitTest', which is now using an 'UnconfinedTestDispatcher' to
instantiate this rule.

In addition to that a new 'test { ... }' utility function is added to
'BaseUnitTest', which will effectively substitutes the 'CoroutineUtils'
such related utility function.

------------------------------------------------------------------------

TODO:
- advanceTimeBy
- pauseDispatcher
- resumeDispatcher

------------------------------------------------------------------------

This is done to match the setup on WCAndroid so that to start making
progress on testing consistency across repos.