Change shouldPublishImmediately to match web logic
After testing web behavior it seem like web shows "Immediately" publish date
for posts that have the same created and modified dates, when they are in the
past.

This commit changes the app behavior to closely match it in the prepublishing
nudge dialog as well as in the actual publish logic from Editor and from the
Posts List.

This change also means that what you see in the prepublishing dialog as publish
date will actually be the publish date, which was not happening before in all
cases.

Caveats: editing posts on the app still set created and modified date to
different dates, so any edited posts will show as "backdated". Manually
setting "Immediately" in prepublishing nudge dialog will also show as
"backdated" as created and modified dates will be different.