Change shouldPublishImmediately to work with publish dialog
Fixes the caveat of the previous commit of manually setting "Immediately" in
prepublishing nudge dialog showing as "backdated" as created and modified dates
will be different. This change adds a "grace period" to avoid false "immediate"
negatives, by allowing posts with different created and modified dates to be
considered "immediate" publish IF the publish date is set to an instant in the
past 30 minutes.

This logic was previously inside the isPublishDateInThePast method, but since
an if-else chain logic was swapped, this makes more sense in this method.