Test: Replace coroutines utils with coroutines test rule test dispatcher
As part of this change the below was done:
- The '@InternalCoroutinesApi' annotation was removed from those tests
(it is no longer needed).
- A new instance of 'UnconfinedTestDispatcher' was added whenever the
'coroutinesTestRule.testDispatcher' could not be accessed
(see 'companion object' access).
- 'test { ... }' was added to some tests in order to access Coroutine
related test scope utilities like 'advanceTimeBy' due to the fact that
'coroutineDispatcher' was no longer accessible as a field.
(see 'SitePreviewViewModelTest' test).
- 'test { ... }' replaced 'runBlockingTest(coroutineContext)' on some
tests due to the fact that 'coroutineContext' was no longer accessible
as a field (see 'SnackbarSequencerConcurrentTest' test).