Test: Fix coroutines related test failure due to field test class init
The 'JetpackMigrationViewModelTest' test suite and some of its tests are
failing because the 'classToTest' object is being initialized via a
field and not during the 'setUp()' phase.

As such, the 'CoroutineTestRule' is not able to run the 'starting(..)'
and 'finished(...)' process that sets and resets the test dispatchers.

------------------------------------------------------------------------

Exception in thread "Test worker" java.lang.IllegalStateException:
Module with the Main dispatcher had failed to initialize. For tests
Dispatchers.setMain from kotlinx-coroutines-test module can be used
java.lang.IllegalStateException: Module with the Main dispatcher had
failed to initialize. For tests Dispatchers.setMain from
kotlinx-coroutines-test module can be used
    at kotlinx.coroutines.internal.MissingMainCoroutineDispatcher
      .missing(MainDispatchers.kt:118)
    ...
Caused by: java.lang.IllegalStateException: The main looper is not
available
    at kotlinx.coroutines.android.AndroidDispatcherFactory
      .createDispatcher(HandlerDispatcher.kt:55)
    ...
    ... 38 more