update unit tests to test the feature flag
In MySiteViewModelTest I found an issue with init being called both in @Before
and in the body of some tests, causing a LiveData to be observed twice and
triggering a verified code twice. A `times(2)` was put in place in the past
to fix the test, and I followed the same approch, but I will try investigating
it further to improve this part.