make FeatureConfig fields in MySiteViewModel lazy
This change makes some fields that store FeatureConfig values on
MySiteViewModel instantitation do that storing lazily. This keeps the class
loading those values only once, but only doing so when actually needed.

Also, this helps with the initialization of the class in unit tests, as the
depedencies that provide those values can be mocked AFTER initiatilization.
With that, I could also move the mocking of such dependencies back to
initSelectedSite in MySiteViewModelTest and avoid calling init twice in
several tests (which was causing some issues in some tests).