add missing ExperimentalCoroutinesApi annotation
Coroutine version was updated and with that the TestDispatcher started being
used in several tests. That is an Experimental Api though, and requires an
annotation of @ExperimentalCoroutinesApi. The PR for coroutine version update
added that change in all but 1 file which is located in a jetpack test source
set, therefore not breaking CI builds: LoginPrologueViewModelTest.

Even though this doesn't break builds (because they compile and run tests on
the WordPress variant, not the Jetpack variant) anyone actively developing
for the Jetpack variant will encounter problems when trying to run unit
tests, as they don't compile.

This change adds the @ExperimentalCoroutinesApi to the file in the jetpack
source set, fixing the COMPILATION issue. Some tests still FAIL when running
the complete unit test suite in the Jetpack variant, but that is expected for
now, as they check specific WordPress assertions.